Clean app.js
